backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pymongo==4.6.1
langchain==0.1.5
openai==1.12.0
pinecone-client==3.0.2
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
pytest==8.0.0
httpx==0.26.0
celery==5.3.6
redis==5.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
email-validator==2.1.0.post1
aiohttp==3.9.3
tenacity==8.2.3
slowapi==0.1.8
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
structlog==24.1.0
gunicorn==21.2.0
uvicorn[standard]==0.27.1
